public thumbnails for album pages too.
