+++ src/pages/portfolio/Collection.tsx
@@ -1,76 +1,17 @@
-import { ReactElement, useEffect, useState } from "react";
+import { useState } from "react";
 import { NavLink, useParams } from "react-router-dom";
-import { CircularProgress } from "@mui/material";
 
-import { getAlbum, getPhoto, Image } from "../../api/rex-api/fetchPhotography";
+import {
+  ALBUMS,
+  getPhoto,
+  getThumbnailURL,
+} from "../../api/rex-api/fetchPhotography";
 
 import SectionTitle from "../../components/SectionTitle";
 import { Thumbnail, ThumbnailGrid } from "../../containers/Thumbnail";
 import { ImageModal } from "../../containers/Modal";
 
-import data from "../../assets/albums.json";
-
 import styles from "./Collection.module.css";
-
-const COLLECTIONS: { [key: string]: Collection } = data;
-
-interface Collection {
-  name: string;
-  thumbnail: {
-    album_name: string;
-    image_name: string;
-  };
-  albums: {
-    name: string;
-    film_stock: string;
-  }[];
-}
-
-const ThumbnailSection = ({
-  album_name,
-  handleOpen,
-  film_stock,
-}: {
-  album_name: string;
-  handleOpen: (album: string, path: string) => void;
-  film_stock?: string;
-}) => {
-  const [albumSection, setAlbumSection] = useState<ReactElement>();
-  const [loaded, setLoaded] = useState<boolean>(false);
-  const [fetchError, setFetchError] = useState<boolean>(false);
-
-  useEffect(() => {
-    getAlbum(album_name)
-      .then((album) => {
-        setAlbumSection(
-          <ThumbnailGrid>
-            {album.images.map((image: Image) => (
-              <Thumbnail
-                key={`${album_name}/${image.path}`}
-                onClick={() => handleOpen(album.name, image.path)}
-                imageSrc={image.url}
-              />
-            ))}
-          </ThumbnailGrid>
-        );
-        setLoaded(true);
-      })
-      .catch(() => {
-        setFetchError(true);
-        setLoaded(true);
-      });
-  }, [album_name, handleOpen]);
-
-  return (
-    <>
-      {film_stock && <h3>{film_stock}</h3>}
-      {fetchError && (
-        <p style={{ color: "red" }}>ERROR RETRIEVING ALBUM: '{album_name}'</p>
-      )}
-      {loaded ? albumSection : <CircularProgress color="inherit" />}
-    </>
-  );
-};
 
 const Collection = () => {
   const { collectionId } = useParams();
@@ -78,20 +19,18 @@ const Collection = () => {
   const [openModal, setOpenModal] = useState<boolean>(false);
   const [imageSrc, setImageSrc] = useState<string>("");
 
-  const handleOpen = (album: string, path: string) => {
-    const split = path.split("/");
-    const name = split[split.length - 1];
-    getPhoto(album, name).then((image) => {
+  const handleOpen = (image_name: string) => {
+    getPhoto(image_name).then((image) => {
       setImageSrc(image.url);
       setOpenModal(true);
     });
   };
 
-  if (!(collectionId! in COLLECTIONS)) {
+  if (!(collectionId! in ALBUMS) || collectionId === undefined) {
     return <h2>NO SUCH COLLECTION...</h2>;
   }
 
-  const collection = COLLECTIONS[collectionId!];
+  const album = ALBUMS[collectionId];
 
   return (
     <>
@@ -99,16 +38,17 @@ const Collection = () => {
         <NavLink to="../" className={styles.Back}>
           {"<- Back"}
         </NavLink>
-        <SectionTitle title={collection.name} />
+        <SectionTitle title={album.name} />
       </div>
-      {collection.albums.map((a) => (
-        <ThumbnailSection
-          key={a.name}
-          album_name={a.name}
-          handleOpen={handleOpen}
-          film_stock={a.film_stock}
-        />
-      ))}
+      <ThumbnailGrid>
+        {Object.entries(album.photos).map(([name]) => (
+          <Thumbnail
+            key={name}
+            onClick={() => handleOpen(name)}
+            imageSrc={getThumbnailURL(name)}
+          />
+        ))}
+      </ThumbnailGrid>
 
       <ImageModal
         isOpen={openModal}
